requirements.txt
Django==3.2
django-cors-headers==3.7.0
django-filter==21.1
djangorestframework==3.12.2
djangorestframework-simplejwt==5.2.2
pytest==6.2.5
pytest-django==4.5.2
psycopg2-binary==2.8
mixer==7.2.2
Pillow==9.4.0